+++ src/Microsoft.Extensions.Configuration.Binder/Microsoft.Extensions.Configuration.Binder.csproj
@@ -4,7 +4,7 @@
 
   <PropertyGroup>
     <Description>Functionality to bind an object to data in configuration providers for Microsoft.Extensions.Configuration.</Description>
-    <TargetFramework>netstandard1.1</TargetFramework>
+    <TargetFramework>netstandard2.0</TargetFramework>
     <GenerateDocumentationFile>true</GenerateDocumentationFile>
     <PackageTags>configuration</PackageTags>
     <EnableApiCheck>false</EnableApiCheck>
@@ -12,10 +12,6 @@
 
   <ItemGroup>
     <ProjectReference Include="..\Microsoft.Extensions.Configuration\Microsoft.Extensions.Configuration.csproj" />
-  </ItemGroup>
-
-  <ItemGroup Condition=" '$(TargetFramework)' == 'netstandard1.1' ">
-    <PackageReference Include="System.ComponentModel.TypeConverter" Version="$(CoreFxVersion)" />
   </ItemGroup>
 
 </Project>
