+++ src/Microsoft.Extensions.Configuration.DockerSecrets/Microsoft.Extensions.Configuration.DockerSecrets.csproj
@@ -4,7 +4,7 @@
 
   <PropertyGroup>
     <Description>Configuration providers that work with docker secrets for Microsoft.Extensions.Configuration.</Description>
-    <TargetFramework>netstandard1.3</TargetFramework>
+    <TargetFramework>netstandard2.0</TargetFramework>
     <GenerateDocumentationFile>true</GenerateDocumentationFile>
     <PackageTags>configuration;docker</PackageTags>
     <EnableApiCheck>false</EnableApiCheck>
@@ -12,11 +12,10 @@
 
   <ItemGroup>
     <ProjectReference Include="..\Microsoft.Extensions.Configuration\Microsoft.Extensions.Configuration.csproj" />
-    <PackageReference Include="Microsoft.Extensions.FileProviders.Physical" Version="$(AspNetCoreVersion)" />
   </ItemGroup>
 
-  <ItemGroup Condition=" '$(TargetFramework)' == 'netstandard1.3' ">
-    <PackageReference Include="System.Threading.Thread" Version="$(CoreFxVersion)" />
+  <ItemGroup>
+    <PackageReference Include="Microsoft.Extensions.FileProviders.Physical" Version="$(AspNetCoreVersion)" />
   </ItemGroup>
 
 </Project>
