+++ src/Microsoft.Extensions.Configuration.EnvironmentVariables/Microsoft.Extensions.Configuration.EnvironmentVariables.csproj
@@ -4,7 +4,7 @@
 
   <PropertyGroup>
     <Description>Environment variables configuration provider implementation for Microsoft.Extensions.Configuration.</Description>
-    <TargetFramework>netstandard1.3</TargetFramework>
+    <TargetFramework>netstandard2.0</TargetFramework>
     <GenerateDocumentationFile>true</GenerateDocumentationFile>
     <PackageTags>configuration;envvar;environmentvariable</PackageTags>
   </PropertyGroup>
