--- src/Microsoft.Extensions.Configuration.EnvironmentVariables/baseline.net45.json
@@ -1,146 +1,0 @@
-{
-  "AssemblyIdentity": "Microsoft.Extensions.Configuration.EnvironmentVariables, Version=1.0.0.0, Culture=neutral, PublicKeyToken=adb9793829ddae60",
-  "Types": [
-    {
-      "Name": "Microsoft.Extensions.Configuration.EnvironmentVariablesExtensions",
-      "Visibility": "Public",
-      "Kind": "Class",
-      "Abstract": true,
-      "Static": true,
-      "Sealed": true,
-      "ImplementedInterfaces": [],
-      "Members": [
-        {
-          "Kind": "Method",
-          "Name": "AddEnvironmentVariables",
-          "Parameters": [
-            {
-              "Name": "configurationBuilder",
-              "Type": "Microsoft.Extensions.Configuration.IConfigurationBuilder"
-            }
-          ],
-          "ReturnType": "Microsoft.Extensions.Configuration.IConfigurationBuilder",
-          "Static": true,
-          "Extension": true,
-          "Visibility": "Public",
-          "GenericParameter": []
-        },
-        {
-          "Kind": "Method",
-          "Name": "AddEnvironmentVariables",
-          "Parameters": [
-            {
-              "Name": "configurationBuilder",
-              "Type": "Microsoft.Extensions.Configuration.IConfigurationBuilder"
-            },
-            {
-              "Name": "prefix",
-              "Type": "System.String"
-            }
-          ],
-          "ReturnType": "Microsoft.Extensions.Configuration.IConfigurationBuilder",
-          "Static": true,
-          "Extension": true,
-          "Visibility": "Public",
-          "GenericParameter": []
-        }
-      ],
-      "GenericParameters": []
-    },
-    {
-      "Name": "Microsoft.Extensions.Configuration.EnvironmentVariables.EnvironmentVariablesConfigurationProvider",
-      "Visibility": "Public",
-      "Kind": "Class",
-      "BaseType": "Microsoft.Extensions.Configuration.ConfigurationProvider",
-      "ImplementedInterfaces": [],
-      "Members": [
-        {
-          "Kind": "Method",
-          "Name": "Load",
-          "Parameters": [],
-          "ReturnType": "System.Void",
-          "Virtual": true,
-          "Override": true,
-          "ImplementedInterface": "Microsoft.Extensions.Configuration.IConfigurationProvider",
-          "Visibility": "Public",
-          "GenericParameter": []
-        },
-        {
-          "Kind": "Constructor",
-          "Name": ".ctor",
-          "Parameters": [],
-          "Visibility": "Public",
-          "GenericParameter": []
-        },
-        {
-          "Kind": "Constructor",
-          "Name": ".ctor",
-          "Parameters": [
-            {
-              "Name": "prefix",
-              "Type": "System.String"
-            }
-          ],
-          "Visibility": "Public",
-          "GenericParameter": []
-        }
-      ],
-      "GenericParameters": []
-    },
-    {
-      "Name": "Microsoft.Extensions.Configuration.EnvironmentVariables.EnvironmentVariablesConfigurationSource",
-      "Visibility": "Public",
-      "Kind": "Class",
-      "ImplementedInterfaces": [
-        "Microsoft.Extensions.Configuration.IConfigurationSource"
-      ],
-      "Members": [
-        {
-          "Kind": "Method",
-          "Name": "get_Prefix",
-          "Parameters": [],
-          "ReturnType": "System.String",
-          "Visibility": "Public",
-          "GenericParameter": []
-        },
-        {
-          "Kind": "Method",
-          "Name": "set_Prefix",
-          "Parameters": [
-            {
-              "Name": "value",
-              "Type": "System.String"
-            }
-          ],
-          "ReturnType": "System.Void",
-          "Visibility": "Public",
-          "GenericParameter": []
-        },
-        {
-          "Kind": "Method",
-          "Name": "Build",
-          "Parameters": [
-            {
-              "Name": "builder",
-              "Type": "Microsoft.Extensions.Configuration.IConfigurationBuilder"
-            }
-          ],
-          "ReturnType": "Microsoft.Extensions.Configuration.IConfigurationProvider",
-          "Sealed": true,
-          "Virtual": true,
-          "ImplementedInterface": "Microsoft.Extensions.Configuration.IConfigurationSource",
-          "Visibility": "Public",
-          "GenericParameter": []
-        },
-        {
-          "Kind": "Constructor",
-          "Name": ".ctor",
-          "Parameters": [],
-          "Visibility": "Public",
-          "GenericParameter": []
-        }
-      ],
-      "GenericParameters": []
-    }
-  ]
-}
