+++ src/Microsoft.Extensions.Configuration.Ini/Microsoft.Extensions.Configuration.Ini.csproj
@@ -4,7 +4,7 @@
 
   <PropertyGroup>
     <Description>INI configuration provider implementation for Microsoft.Extensions.Configuration.</Description>
-    <TargetFramework>netstandard1.3</TargetFramework>
+    <TargetFramework>netstandard2.0</TargetFramework>
     <GenerateDocumentationFile>true</GenerateDocumentationFile>
     <PackageTags>configuration;ini</PackageTags>
   </PropertyGroup>
