--- src/Microsoft.Extensions.Configuration.Ini/baseline.net45.json
@@ -1,187 +1,0 @@
-{
-  "AssemblyIdentity": "Microsoft.Extensions.Configuration.Ini, Version=1.0.0.0, Culture=neutral, PublicKeyToken=adb9793829ddae60",
-  "Types": [
-    {
-      "Name": "Microsoft.Extensions.Configuration.IniConfigurationExtensions",
-      "Visibility": "Public",
-      "Kind": "Class",
-      "Abstract": true,
-      "Static": true,
-      "Sealed": true,
-      "ImplementedInterfaces": [],
-      "Members": [
-        {
-          "Kind": "Method",
-          "Name": "AddIniFile",
-          "Parameters": [
-            {
-              "Name": "builder",
-              "Type": "Microsoft.Extensions.Configuration.IConfigurationBuilder"
-            },
-            {
-              "Name": "path",
-              "Type": "System.String"
-            }
-          ],
-          "ReturnType": "Microsoft.Extensions.Configuration.IConfigurationBuilder",
-          "Static": true,
-          "Extension": true,
-          "Visibility": "Public",
-          "GenericParameter": []
-        },
-        {
-          "Kind": "Method",
-          "Name": "AddIniFile",
-          "Parameters": [
-            {
-              "Name": "builder",
-              "Type": "Microsoft.Extensions.Configuration.IConfigurationBuilder"
-            },
-            {
-              "Name": "path",
-              "Type": "System.String"
-            },
-            {
-              "Name": "optional",
-              "Type": "System.Boolean"
-            }
-          ],
-          "ReturnType": "Microsoft.Extensions.Configuration.IConfigurationBuilder",
-          "Static": true,
-          "Extension": true,
-          "Visibility": "Public",
-          "GenericParameter": []
-        },
-        {
-          "Kind": "Method",
-          "Name": "AddIniFile",
-          "Parameters": [
-            {
-              "Name": "builder",
-              "Type": "Microsoft.Extensions.Configuration.IConfigurationBuilder"
-            },
-            {
-              "Name": "path",
-              "Type": "System.String"
-            },
-            {
-              "Name": "optional",
-              "Type": "System.Boolean"
-            },
-            {
-              "Name": "reloadOnChange",
-              "Type": "System.Boolean"
-            }
-          ],
-          "ReturnType": "Microsoft.Extensions.Configuration.IConfigurationBuilder",
-          "Static": true,
-          "Extension": true,
-          "Visibility": "Public",
-          "GenericParameter": []
-        },
-        {
-          "Kind": "Method",
-          "Name": "AddIniFile",
-          "Parameters": [
-            {
-              "Name": "builder",
-              "Type": "Microsoft.Extensions.Configuration.IConfigurationBuilder"
-            },
-            {
-              "Name": "provider",
-              "Type": "Microsoft.Extensions.FileProviders.IFileProvider"
-            },
-            {
-              "Name": "path",
-              "Type": "System.String"
-            },
-            {
-              "Name": "optional",
-              "Type": "System.Boolean"
-            },
-            {
-              "Name": "reloadOnChange",
-              "Type": "System.Boolean"
-            }
-          ],
-          "ReturnType": "Microsoft.Extensions.Configuration.IConfigurationBuilder",
-          "Static": true,
-          "Extension": true,
-          "Visibility": "Public",
-          "GenericParameter": []
-        }
-      ],
-      "GenericParameters": []
-    },
-    {
-      "Name": "Microsoft.Extensions.Configuration.Ini.IniConfigurationProvider",
-      "Visibility": "Public",
-      "Kind": "Class",
-      "BaseType": "Microsoft.Extensions.Configuration.FileConfigurationProvider",
-      "ImplementedInterfaces": [],
-      "Members": [
-        {
-          "Kind": "Method",
-          "Name": "Load",
-          "Parameters": [
-            {
-              "Name": "stream",
-              "Type": "System.IO.Stream"
-            }
-          ],
-          "ReturnType": "System.Void",
-          "Virtual": true,
-          "Override": true,
-          "Visibility": "Public",
-          "GenericParameter": []
-        },
-        {
-          "Kind": "Constructor",
-          "Name": ".ctor",
-          "Parameters": [
-            {
-              "Name": "source",
-              "Type": "Microsoft.Extensions.Configuration.Ini.IniConfigurationSource"
-            }
-          ],
-          "Visibility": "Public",
-          "GenericParameter": []
-        }
-      ],
-      "GenericParameters": []
-    },
-    {
-      "Name": "Microsoft.Extensions.Configuration.Ini.IniConfigurationSource",
-      "Visibility": "Public",
-      "Kind": "Class",
-      "BaseType": "Microsoft.Extensions.Configuration.FileConfigurationSource",
-      "ImplementedInterfaces": [],
-      "Members": [
-        {
-          "Kind": "Method",
-          "Name": "Build",
-          "Parameters": [
-            {
-              "Name": "builder",
-              "Type": "Microsoft.Extensions.Configuration.IConfigurationBuilder"
-            }
-          ],
-          "ReturnType": "Microsoft.Extensions.Configuration.IConfigurationProvider",
-          "Virtual": true,
-          "Override": true,
-          "ImplementedInterface": "Microsoft.Extensions.Configuration.IConfigurationSource",
-          "Visibility": "Public",
-          "GenericParameter": []
-        },
-        {
-          "Kind": "Constructor",
-          "Name": ".ctor",
-          "Parameters": [],
-          "Visibility": "Public",
-          "GenericParameter": []
-        }
-      ],
-      "GenericParameters": []
-    }
-  ]
-}
