+++ src/Microsoft.Extensions.Configuration.UserSecrets/Microsoft.Extensions.Configuration.UserSecrets.csproj
@@ -4,7 +4,7 @@
 
   <PropertyGroup>
     <Description>User secrets configuration provider implementation for Microsoft.Extensions.Configuration.</Description>
-    <TargetFramework>netstandard1.3</TargetFramework>
+    <TargetFramework>netstandard2.0</TargetFramework>
     <GenerateDocumentationFile>true</GenerateDocumentationFile>
     <PackageTags>configuration;secrets;usersecrets</PackageTags>
   </PropertyGroup>
