+++ src/Microsoft.Extensions.Configuration.Xml/Microsoft.Extensions.Configuration.Xml.csproj
@@ -4,7 +4,7 @@
 
   <PropertyGroup>
     <Description>XML configuration provider implementation for Microsoft.Extensions.Configuration.</Description>
-    <TargetFrameworks>net461;netstandard1.3</TargetFrameworks>
+    <TargetFramework>netstandard2.0</TargetFramework>
     <GenerateDocumentationFile>true</GenerateDocumentationFile>
     <PackageTags>configuration;xml</PackageTags>
   </PropertyGroup>
@@ -14,14 +14,8 @@
     <ProjectReference Include="..\Microsoft.Extensions.Configuration.FileExtensions\Microsoft.Extensions.Configuration.FileExtensions.csproj" />
   </ItemGroup>
 
-  <ItemGroup Condition=" '$(TargetFramework)' == 'net461' ">
-    <Reference Include="System.Security" />
-    <Reference Include="System.Xml" />
-  </ItemGroup>
-
-  <ItemGroup Condition=" '$(TargetFramework)' == 'netstandard1.3' ">
-    <PackageReference Include="System.Xml.XmlDocument" Version="$(CoreFxVersion)" />
-    <PackageReference Include="System.Xml.XmlSerializer" Version="$(CoreFxVersion)" />
+  <ItemGroup>
+    <PackageReference Include="System.Security.Cryptography.Xml" Version="$(CoreFxVersion)" />
   </ItemGroup>
 
 </Project>
