+++ src/Microsoft.Extensions.Configuration.Xml/Resources.resx
@@ -117,9 +117,6 @@
   <resheader name="writer">
     <value>System.Resources.ResXResourceWriter, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
   </resheader>
-  <data name="Error_EncryptedXmlNotSupported" xml:space="preserve">
-    <value>Encrypted XML is not supported on this platform.</value>
-  </data>
   <data name="Error_InvalidFilePath" xml:space="preserve">
     <value>File path must be a non-empty string.</value>
   </data>
