--- src/Microsoft.Extensions.Configuration.Xml/baseline.netframework.json
@@ -1,244 +1,0 @@
-{
-  "AssemblyIdentity": "Microsoft.Extensions.Configuration.Xml, Version=1.0.0.0, Culture=neutral, PublicKeyToken=adb9793829ddae60",
-  "Types": [
-    {
-      "Name": "Microsoft.Extensions.Configuration.XmlConfigurationExtensions",
-      "Visibility": "Public",
-      "Kind": "Class",
-      "Abstract": true,
-      "Static": true,
-      "Sealed": true,
-      "ImplementedInterfaces": [],
-      "Members": [
-        {
-          "Kind": "Method",
-          "Name": "AddXmlFile",
-          "Parameters": [
-            {
-              "Name": "builder",
-              "Type": "Microsoft.Extensions.Configuration.IConfigurationBuilder"
-            },
-            {
-              "Name": "path",
-              "Type": "System.String"
-            }
-          ],
-          "ReturnType": "Microsoft.Extensions.Configuration.IConfigurationBuilder",
-          "Static": true,
-          "Extension": true,
-          "Visibility": "Public",
-          "GenericParameter": []
-        },
-        {
-          "Kind": "Method",
-          "Name": "AddXmlFile",
-          "Parameters": [
-            {
-              "Name": "builder",
-              "Type": "Microsoft.Extensions.Configuration.IConfigurationBuilder"
-            },
-            {
-              "Name": "path",
-              "Type": "System.String"
-            },
-            {
-              "Name": "optional",
-              "Type": "System.Boolean"
-            }
-          ],
-          "ReturnType": "Microsoft.Extensions.Configuration.IConfigurationBuilder",
-          "Static": true,
-          "Extension": true,
-          "Visibility": "Public",
-          "GenericParameter": []
-        },
-        {
-          "Kind": "Method",
-          "Name": "AddXmlFile",
-          "Parameters": [
-            {
-              "Name": "builder",
-              "Type": "Microsoft.Extensions.Configuration.IConfigurationBuilder"
-            },
-            {
-              "Name": "path",
-              "Type": "System.String"
-            },
-            {
-              "Name": "optional",
-              "Type": "System.Boolean"
-            },
-            {
-              "Name": "reloadOnChange",
-              "Type": "System.Boolean"
-            }
-          ],
-          "ReturnType": "Microsoft.Extensions.Configuration.IConfigurationBuilder",
-          "Static": true,
-          "Extension": true,
-          "Visibility": "Public",
-          "GenericParameter": []
-        },
-        {
-          "Kind": "Method",
-          "Name": "AddXmlFile",
-          "Parameters": [
-            {
-              "Name": "builder",
-              "Type": "Microsoft.Extensions.Configuration.IConfigurationBuilder"
-            },
-            {
-              "Name": "provider",
-              "Type": "Microsoft.Extensions.FileProviders.IFileProvider"
-            },
-            {
-              "Name": "path",
-              "Type": "System.String"
-            },
-            {
-              "Name": "optional",
-              "Type": "System.Boolean"
-            },
-            {
-              "Name": "reloadOnChange",
-              "Type": "System.Boolean"
-            }
-          ],
-          "ReturnType": "Microsoft.Extensions.Configuration.IConfigurationBuilder",
-          "Static": true,
-          "Extension": true,
-          "Visibility": "Public",
-          "GenericParameter": []
-        }
-      ],
-      "GenericParameters": []
-    },
-    {
-      "Name": "Microsoft.Extensions.Configuration.Xml.XmlConfigurationProvider",
-      "Visibility": "Public",
-      "Kind": "Class",
-      "BaseType": "Microsoft.Extensions.Configuration.FileConfigurationProvider",
-      "ImplementedInterfaces": [],
-      "Members": [
-        {
-          "Kind": "Method",
-          "Name": "Load",
-          "Parameters": [
-            {
-              "Name": "stream",
-              "Type": "System.IO.Stream"
-            }
-          ],
-          "ReturnType": "System.Void",
-          "Virtual": true,
-          "Override": true,
-          "Visibility": "Public",
-          "GenericParameter": []
-        },
-        {
-          "Kind": "Constructor",
-          "Name": ".ctor",
-          "Parameters": [
-            {
-              "Name": "source",
-              "Type": "Microsoft.Extensions.Configuration.Xml.XmlConfigurationSource"
-            }
-          ],
-          "Visibility": "Public",
-          "GenericParameter": []
-        }
-      ],
-      "GenericParameters": []
-    },
-    {
-      "Name": "Microsoft.Extensions.Configuration.Xml.XmlConfigurationSource",
-      "Visibility": "Public",
-      "Kind": "Class",
-      "BaseType": "Microsoft.Extensions.Configuration.FileConfigurationSource",
-      "ImplementedInterfaces": [],
-      "Members": [
-        {
-          "Kind": "Method",
-          "Name": "Build",
-          "Parameters": [
-            {
-              "Name": "builder",
-              "Type": "Microsoft.Extensions.Configuration.IConfigurationBuilder"
-            }
-          ],
-          "ReturnType": "Microsoft.Extensions.Configuration.IConfigurationProvider",
-          "Virtual": true,
-          "Override": true,
-          "ImplementedInterface": "Microsoft.Extensions.Configuration.IConfigurationSource",
-          "Visibility": "Public",
-          "GenericParameter": []
-        },
-        {
-          "Kind": "Constructor",
-          "Name": ".ctor",
-          "Parameters": [],
-          "Visibility": "Public",
-          "GenericParameter": []
-        }
-      ],
-      "GenericParameters": []
-    },
-    {
-      "Name": "Microsoft.Extensions.Configuration.Xml.XmlDocumentDecryptor",
-      "Visibility": "Public",
-      "Kind": "Class",
-      "ImplementedInterfaces": [],
-      "Members": [
-        {
-          "Kind": "Method",
-          "Name": "CreateDecryptingXmlReader",
-          "Parameters": [
-            {
-              "Name": "input",
-              "Type": "System.IO.Stream"
-            },
-            {
-              "Name": "settings",
-              "Type": "System.Xml.XmlReaderSettings"
-            }
-          ],
-          "ReturnType": "System.Xml.XmlReader",
-          "Visibility": "Public",
-          "GenericParameter": []
-        },
-        {
-          "Kind": "Method",
-          "Name": "DecryptDocumentAndCreateXmlReader",
-          "Parameters": [
-            {
-              "Name": "document",
-              "Type": "System.Xml.XmlDocument"
-            }
-          ],
-          "ReturnType": "System.Xml.XmlReader",
-          "Virtual": true,
-          "Visibility": "Protected",
-          "GenericParameter": []
-        },
-        {
-          "Kind": "Constructor",
-          "Name": ".ctor",
-          "Parameters": [],
-          "Visibility": "Protected",
-          "GenericParameter": []
-        },
-        {
-          "Kind": "Field",
-          "Name": "Instance",
-          "Parameters": [],
-          "ReturnType": "Microsoft.Extensions.Configuration.Xml.XmlDocumentDecryptor",
-          "Static": true,
-          "ReadOnly": true,
-          "Visibility": "Public",
-          "GenericParameter": []
-        }
-      ],
-      "GenericParameters": []
-    }
-  ]
-}
