+++ src/Microsoft.Extensions.Configuration/Microsoft.Extensions.Configuration.csproj
@@ -4,7 +4,7 @@
 
   <PropertyGroup>
     <Description>Implementation of key-value pair based configuration for Microsoft.Extensions.Configuration. Includes the memory configuration provider.</Description>
-    <TargetFramework>netstandard1.1</TargetFramework>
+    <TargetFramework>netstandard2.0</TargetFramework>
     <GenerateDocumentationFile>true</GenerateDocumentationFile>
     <PackageTags>configuration</PackageTags>
   </PropertyGroup>
