+++ test/Microsoft.Extensions.Configuration.Xml.Test/XmlConfigurationTest.cs
@@ -3,16 +3,13 @@
 
 using System;
 using System.IO;
-using System.Xml;
-using Microsoft.AspNetCore.Testing;
-using Microsoft.Extensions.Configuration.Test;
-using Xunit;
-
-#if NET461
 using System.Security.Cryptography;
 using System.Security.Cryptography.Xml;
+using System.Xml;
+using Microsoft.AspNetCore.Testing;
 using Microsoft.AspNetCore.Testing.xunit;
-#endif
+using Microsoft.Extensions.Configuration.Test;
+using Xunit;
 
 namespace Microsoft.Extensions.Configuration.Xml.Test
 {
@@ -401,48 +398,6 @@ namespace Microsoft.Extensions.Configuration.Xml.Test
             var config = new ConfigurationBuilder().AddXmlFile("NotExistingConfig.xml", optional: true).Build();
         }
 
-#if NETCOREAPP2_0
-        [Fact]
-        public void LoadKeyValuePairsFromValidEncryptedXml_ThrowsPlatformNotSupported()
-        {
-            var xml = @"
-                <settings>
-                    <Data.Setting>
-                        <DefaultConnection>
-                            <Connection.String>Test.Connection.String</Connection.String>
-                            <Provider>SqlClient</Provider>
-                        </DefaultConnection>
-                        <Inventory>
-                            <EncryptedData Type=""http://www.w3.org/2001/04/xmlenc#Element"" xmlns=""http://www.w3.org/2001/04/xmlenc#"">
-                            <EncryptionMethod Algorithm=""http://www.w3.org/2001/04/xmlenc#aes256-cbc"" />
-                            <KeyInfo xmlns=""http://www.w3.org/2000/09/xmldsig#"">
-                                <EncryptedKey xmlns=""http://www.w3.org/2001/04/xmlenc#"">
-                                <EncryptionMethod Algorithm=""http://www.w3.org/2001/04/xmlenc#kw-aes256"" />
-                                <KeyInfo xmlns=""http://www.w3.org/2000/09/xmldsig#"">
-                                    <KeyName>myKey</KeyName>
-                                </KeyInfo>
-                                <CipherData>
-                                    <CipherValue>b0dxJI/o00vZgTNOJ6wUt0/6wCKWlQANAYE8cBsEzok4LQma7ErEnA==</CipherValue>
-                                </CipherData>
-                                </EncryptedKey>
-                            </KeyInfo>
-                            <CipherData>
-                                <CipherValue>iXzecb+Cha80LLrl4zON3o7HfpRc0NxlJsnBb6zbKFa1HqtNhy2VrTnrEsZUViBWRkRbl+MCix7TiaIs4NtLijNU5Ob8Ez3vcD4T/QcmPywBYJDJhj1OUUeJSKH+icjg</CipherValue>
-                            </CipherData>
-                            </EncryptedData>
-                            <Provider>MySql</Provider>
-                        </Inventory>
-                    </Data.Setting>
-                </settings>";
-
-            // Arrange
-            var xmlConfigSrc = new XmlConfigurationProvider(new XmlConfigurationSource());
-
-            // Act & assert
-            var ex = Assert.Throws<PlatformNotSupportedException>(() => xmlConfigSrc.Load(TestStreamHelpers.StringToStream(xml)));
-            Assert.Equal(Resources.Error_EncryptedXmlNotSupported, ex.Message);
-        }
-#elif NET461
         [ConditionalFact]
         [FrameworkSkipCondition(RuntimeFrameworks.Mono)]
         public void LoadKeyValuePairsFromValidEncryptedXml()
@@ -479,7 +434,7 @@ namespace Microsoft.Extensions.Configuration.Xml.Test
 
             // Arrange
             var xmlConfigSrc = new XmlConfigurationProvider(new XmlConfigurationSource());
-            xmlConfigSrc.Decryptor = new EncryptedXmlDocumentDecryptor(doc =>
+            xmlConfigSrc.Decryptor = new XmlDocumentDecryptor(doc =>
             {
                 var innerEncryptedXml = new EncryptedXml(doc);
                 innerEncryptedXml.AddKeyNameMapping("myKey", aes);
@@ -495,8 +450,5 @@ namespace Microsoft.Extensions.Configuration.Xml.Test
             Assert.Equal("AnotherTestConnectionString", xmlConfigSrc.Get("data.setting:inventory:connectionstring"));
             Assert.Equal("MySql", xmlConfigSrc.Get("Data.setting:Inventory:Provider"));
         }
-#else
-#error Target framework needs to be updated
-#endif
     }
 }
